Client-common separation is finished
