+++ src/engine/GameStates.js
@@ -1,0 +1,5 @@
+export default {
+    start: Symbol('start'),
+    play: Symbol('play'),
+    pause: Symbol('pause'),
+};
