+++ src/engine/util.js
@@ -15,29 +15,3 @@ export function animate(dx, startTime, currentTime, speed, looped = false) {
 
     return dx * t;
 }
-
-export function animateObject(obj, { newX, newY, map, speed, time }) {
-    if (time) {
-        const [newX, newY] = [
-            obj.toX + animate(obj.deltaX, obj.motionStartTime, time, obj.speed) - obj.deltaX,
-            obj.toY + animate(obj.deltaY, obj.motionStartTime, time, obj.speed) - obj.deltaY,
-        ];
-
-        if (newX === obj.toX && newY === obj.toY) {
-            obj.speed = 0;
-            obj.motionStartTime = 0;
-            if (obj['trigger']) obj.trigger('animation-stopped');
-        }
-
-        obj.x = newX;
-        obj.y = newY;
-    } else {
-        obj.motionStartTime = map.engine.lastRenderTime;
-        obj.speed = speed;
-        obj.toX = newX;
-        obj.toY = newY;
-        obj.deltaX = newX - obj.x;
-        obj.deltaY = newY - obj.y;
-        if (obj['trigger']) obj.trigger('animation-started');
-    }
-}
